+++ lib/my/my_nbr_len.c
@@ -7,7 +7,7 @@
 
 #include <my.h>
 
-int my_nbr_len(int nb)
+int my_nbr_len(long nb)
 {
     int len = 1;
